20240808 - feat: todoApp 컴포넌트화 TodoInput, TodoList, TodoItem 컴포넌트 추가 Data타입을 index.d.ts에 추가하여 전역에서 사용할 수 있게 추가
