+++ types/index.d.ts
@@ -1,0 +1,5 @@
+interface Data {
+  id: number;
+  todo: string;
+  isDone: boolean;
+}
